\name{ssa}
\alias{ssa}
\docType{data}
\title{
data on health expenditure, per capita income and life expectancy in Sub-Saharan countries
}
\description{
panel data on 
}
\usage{data("ssa")}
\format{
  A panel data frame  on the following variables.
  \describe{
    \item{\code{id}}{list of SSA countries }
    \item{\code{year}}{year}
    \item{\code{HEP}}{health expenditure per capital}
    \item{\code{PCI}}{Per capita income}
    \item{\code{LE}}{Life expectancy}
  }
}
\note{
variables are in natural log
}
\source{
WDI
WHO
}
\keyword{datasets}
